Add a pre-config that logs all new non-local ip connections.
A pre-config file can be enabled by defining a new type in the stack
environment file and an associating it with a yaml definition.
The environment will then overwrite the default config that is set to
OS::Heat::None.
All template files have been updated to heat_template_version:
2015-04-30 to enable overwriting the stack_id of a stack during creation
time.